+++ NEWS.md
@@ -1,5 +1,7 @@
 # fs (development version)
 
+* `dir_create()` now restores the previous umask (#293)
+
 # fs 1.5.0
 
 * The libuv release used by fs was updated to 1.38.1